refactor agnews example
