+++ example/text_cls_AG_NEWS/.gitignore
@@ -1,0 +1,2 @@
+data/
+models/
