+++ example/text_cls_AG_NEWS/.swignore
@@ -1,0 +1,5 @@
+ venv
+.git
+.history
+.vscode
+.venv
