--- example/text_cls_AG_NEWS/requirements.txt
@@ -1,5 +1,0 @@
-numpy
-starwhale
-torch
-torchdata
-torchtext
